+++ src/OpenTelemetry.Instrumentation.SqlClient/SqlClientTraceInstrumentationOptions.cs
@@ -81,21 +81,20 @@ public class SqlClientTraceInstrumentationOptions
     /// cref="SqlClientInstrumentation"/> should parse the DataSource on a
     /// SqlConnection into server name, instance name, and/or port
     /// connection-level attribute tags. Default value: <see
-    /// langword="false"/>.
+    /// langword="true"/>.
     /// </summary>
     /// <remarks>
     /// <para>
     /// <b>EnableConnectionLevelAttributes is supported on all runtimes.</b>
     /// </para>
     /// <para>
-    /// The default behavior is to set the SqlConnection DataSource as the <see cref="SemanticConventions.AttributePeerService"/> tag.
     /// If enabled, SqlConnection DataSource will be parsed and the server name will be sent as the
-    /// <see cref="SemanticConventions.AttributeServerAddress"/> or <see cref="SemanticConventions.AttributeServerSocketAddress"/> tag,
+    /// <see cref="SemanticConventions.AttributeServerAddress"/> tag,
     /// the instance name will be sent as the <see cref="SemanticConventions.AttributeDbMsSqlInstanceName"/> tag,
     /// and the port will be sent as the <see cref="SemanticConventions.AttributeServerPort"/> tag if it is not 1433 (the default port).
     /// </para>
     /// </remarks>
-    public bool EnableConnectionLevelAttributes { get; set; }
+    public bool EnableConnectionLevelAttributes { get; set; } = true;
 
     /// <summary>
     /// Gets or sets an action to enrich an <see cref="Activity"/> with the